requirements.txt
absl-py==1.2.0
addict==2.4.0
asttokens==2.0.8
attributee==0.1.7
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.24.76
botocore==1.27.76
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
dotty-dict==1.3.1
executing==1.0.0
filelock==3.8.0
fire==0.4.0
fonttools==4.37.2
gdown==4.5.1
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.4
importlib-metadata==4.12.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
kiwisolver==1.4.4
lap==0.4.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mmcls==0.23.2
mmcv-full==1.6.1
mmdet==2.25.2
mmengine==0.1.0
mmtrack==0.14.0
model-index==0.1.11
motmetrics==1.2.5
numpy==1.22.4
oauthlib==3.2.1
opencv-python==4.6.0.66
openmim==0.3.1
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.4
pycocotools==2.0.5
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
s3transfer==0.6.0
sahi==0.10.5
scipy==1.7.3
seaborn==0.12.0
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xmltodict==0.13.0
yapf==0.32.0
yolov5==6.2.0
zipp==3.8.1
